move: AdditionalButton from AddButton directory
